feat(frontend): 復元ボタンを押すと元のプレイリストに復元する処理を実装 #138

+++ src/hooks/useSleepTracks.ts
@@ -5,6 +5,7 @@ import useActionSuccessMessageStore from "../store/actionSuccessMessageStore";
 import useTrackMoreMenuStore from "../store/trackMoreMenuStore";
 import usePlaybackStore from "../store/playbackStore";
 import type { ActionType } from "../types/actionType";
+import type { SpotifyTrack, LocalTrack } from "../types/tracksType";
 import { API } from "../api/apis";
 import { STORAGE_KEYS } from "../utils/storageKeys";
 
@@ -88,19 +89,46 @@ const useSleepTracks = () => {
     }
   }
 
-  async function restoreSleepTrack(trackId: string) {
+  async function restoreSleepTrack(trackId: string | undefined, playlistRef: string | undefined) {
     let removedTrack = null;
+    const cached = localStorage.getItem(STORAGE_KEYS.SLEEP_TRACKS);
+    const cachedTracks: (SpotifyTrack | LocalTrack)[] = JSON.parse(cached!);
 
     try {
+      if (!trackId) throw new Error("trackIdが無効");
+      if (!playlistRef) throw new Error("playlistRefが無効");
+
       const response = await fetch(API.deleteSleepSpotifyTracks(trackId), { method: "DELETE" });
 
       if (!response.ok) throw new Error(response.statusText);
 
       removedTrack = await response.json();
+
+      const updateTracks = cachedTracks.filter((track) => track.id !== trackId);
+      setTracks(updateTracks);
+      setQueue(updateTracks);
+      localStorage.setItem(STORAGE_KEYS.SLEEP_TRACKS, JSON.stringify(updateTracks));
     } catch (error) {
       console.error("スリープ曲の復元に失敗:", error);
       showMessage("sleepSpotifyRestoreFailed");
       return;
+    }
+
+    try {
+      const response = await fetch(API.restoreSleepSpotifyTracks(playlistRef), {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(removedTrack),
+      });
+
+      if (!response.ok) throw new Error(response.statusText);
+
+      const { addedTrack } = await response.json();
+    } catch (error) {
+      console.error("スリープ曲の復元に失敗:", error);
+      showMessage("sleepSpotifyRestoreFailed");
     }
   }
